requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
optuna>=3.4.0
streamlit>=1.28.0
plotly>=5.18.0
python-dotenv>=1.0.0
requests>=2.31.0
google-api-python-client>=2.108.0
pytest>=7.4.0
jupyter>=1.0.0
seaborn>=0.13.0
